emailScanner/requirements.txt
# Email Scanner - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
sqlalchemy>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Email handling
# imaplib is part of Python standard library

# AI/ML
requests>=2.31.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database drivers (optional - uncomment if needed)
# psycopg2-binary>=2.9.9  # PostgreSQL
# pymysql>=1.1.0          # MySQL

# Development dependencies (optional)
# pytest>=7.4.3
# pytest-cov>=4.1.0
# black>=23.11.0
# flake8>=6.1.0
